fix(editorial-theme): keep astro:content out of package entry for astro config load
